refactor code + add test for byte offsets
