+++ crates/lsp/src/lib.rs
@@ -6,6 +6,9 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+pub mod build;
+pub mod compiler;
+pub mod lint;
 pub mod lsp;
 pub mod utils;
